pyproject.toml
[tool.poetry]
name = "purgebot"
version = "1.0.0"
description = "A trivial Discord bot for kicking members from servers based on a condition."
authors = ["travis-bittel <travis.bittel2@gmail.com>"]
readme = "README.md"
packages = [
    { include = "bot" },
]

[tool.poetry.dependencies]
python = "3.9"
discord-py = "2.4.0"
boto3 = "1.35.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry-plugin-lambda-build]
package-artifact-path = "bot_lambda.zip"

[tool.poetry.dev-dependencies]
pytest = "8.3.2"

[tool.pytest.ini_options]
minversion = "6.0"
pythonpath = "src"
testpaths = [
    "bot/tests/*"
]
